+++ javascripts/save_load.js
@@ -1,8 +1,8 @@
-function rot(s, i) {
+/*function rot(s, i) {
 	return s.replace(/[a-zA-Z]/g, function (c) {
 		return String.fromCharCode((c <= "Z" ? 90 : 122) >= (c = c.charCodeAt(0) + i) ? c : c - 26);
 	});
-}
+}*/
 
 function getSaveString() {
 	return LZString.compressToBase64(JSON.stringify(player));